SNOW-1825478-PAT-support
