+++ lib/util.js
@@ -4,7 +4,6 @@ const os = require('os');
 const Logger = require('./logger');
 const fs = require('fs');
 const Errors = require('./errors');
-const net = require('net');
 
 /**
  * Note: A simple wrapper around util.inherits() for now, but this might change
@@ -765,18 +764,6 @@ exports.isEmptyObject = (object) => {
 exports.isWindows = function () {
   return os.platform() === 'win32';
 };
-
-
-exports.getFreePort = async function () {
-  return new Promise(res => {
-    const srv = net.createServer();
-    srv.listen(0, () => {
-      const port = srv.address().port;
-      srv.close(() => res(port));
-    });
-  });
-};
-
 
 /**
 * Left strip the specified character from a string.
